requirements.txt
# Discordライブラリ
discord.py
# Minecraftサーバー状態取得ライブラリ
mcstatus
# RCONクライアントライブラリ
mcrcon
